Use plugintimer, fix to random album, zeroconf to module
